isolate contract io to different packages
