+++ op-acceptance-tests/tests/smoke/smoke_test.go
@@ -4,11 +4,11 @@ import (
 	"testing"
 
 	"github.com/ethereum-optimism/optimism/op-devstack/devtest"
-	"github.com/ethereum-optimism/optimism/op-devstack/dsl"
+	"github.com/ethereum-optimism/optimism/op-devstack/dsl/contractdsl"
 	"github.com/ethereum-optimism/optimism/op-devstack/presets"
 	"github.com/ethereum-optimism/optimism/op-service/eth"
-	"github.com/ethereum-optimism/optimism/op-service/txintent"
 	"github.com/ethereum-optimism/optimism/op-service/txintent/bindings"
+	"github.com/ethereum-optimism/optimism/op-service/txintent/contractio"
 	"github.com/ethereum-optimism/optimism/op-service/txplan"
 )
 
@@ -42,48 +42,46 @@ func TestWrapETH(gt *testing.T) {
 	require.NotEqual(alice.Address(), bob.Address())
 
 	// Alice and Bob has zero WETH
-	require.Equal(eth.ZeroWei, dsl.Read(weth.BalanceOf(alice.Address())))
-	require.Equal(eth.ZeroWei, dsl.Read(weth.BalanceOf(bob.Address())))
+	require.Equal(eth.ZeroWei, contractdsl.Read(weth.BalanceOf(alice.Address())))
+	require.Equal(eth.ZeroWei, contractdsl.Read(weth.BalanceOf(bob.Address())))
 
 	// Write: Alice wraps 1 WETH
 	alice.Transfer(*wethAddr, eth.OneEther)
 
 	// Read: Alice has 1 WETH
-	require.Equal(eth.OneEther, dsl.Read(weth.BalanceOf(alice.Address())))
+	require.Equal(eth.OneEther, contractdsl.Read(weth.BalanceOf(alice.Address())))
 	// Read: Bob has 0 WETH
-	require.Equal(eth.ZeroWei, dsl.Read(weth.BalanceOf(bob.Address())))
+	require.Equal(eth.ZeroWei, contractdsl.Read(weth.BalanceOf(bob.Address())))
 
 	// Write: Alice wraps 1 WETH again
 	alice.Transfer(*wethAddr, eth.OneEther)
 
 	// Read: Alice has 2 WETH
-	require.Equal(eth.Ether(2), dsl.Read(weth.BalanceOf(alice.Address())))
+	require.Equal(eth.Ether(2), contractdsl.Read(weth.BalanceOf(alice.Address())))
 	// Read: Bob has 0 WETH
-	require.Equal(eth.ZeroWei, dsl.Read(weth.BalanceOf(bob.Address())))
+	require.Equal(eth.ZeroWei, contractdsl.Read(weth.BalanceOf(bob.Address())))
 
 	// Read not using the DSL. Therefore you need to manually error handle and also set context
-	_, err := txintent.Read(weth.Transfer(bob.Address(), eth.OneEther), t.Ctx())
+	_, err := contractio.Read(weth.Transfer(bob.Address(), eth.OneEther), t.Ctx())
 	// Will revert because tx.sender is not set
 	t.Require().Error(err)
 	// Provide tx.sender using txplan
 	// Success because tx.sender(Alice) has enough WETH
-	require.True(dsl.Read(weth.Transfer(bob.Address(), eth.OneEther), txplan.WithSender(alice.Address())))
+	require.True(contractdsl.Read(weth.Transfer(bob.Address(), eth.OneEther), txplan.WithSender(alice.Address())))
 
 	// Write: Alice sends Bob 1 WETH
-	dsl.Write(alice, weth.Transfer(bob.Address(), eth.OneEther))
+	contractdsl.Write(alice, weth.Transfer(bob.Address(), eth.OneEther))
 
 	// Read: Alice has 1 WETH
-	require.Equal(eth.OneEther, dsl.Read(weth.BalanceOf(alice.Address())))
+	require.Equal(eth.OneEther, contractdsl.Read(weth.BalanceOf(alice.Address())))
 	// Read: Bob has 1 WETH
-	require.Equal(eth.OneEther, dsl.Read(weth.BalanceOf(bob.Address())))
+	require.Equal(eth.OneEther, contractdsl.Read(weth.BalanceOf(bob.Address())))
 
-	// Write not using the DSL. Therefore you need to manually error handle
-	_, err = alice.Write(weth.Transfer(bob.Address(), eth.OneEther))
-	// Success because tx.sender(Alice) has enough WETH
-	require.NoError(err)
+	// Write: Alice sends Bob 1 WETH
+	contractdsl.Write(alice, weth.Transfer(bob.Address(), eth.OneEther))
 
 	// Read: Alice has 0 WETH
-	require.Equal(eth.ZeroWei, dsl.Read(weth.BalanceOf(alice.Address())))
+	require.Equal(eth.ZeroWei, contractdsl.Read(weth.BalanceOf(alice.Address())))
 	// Read: Bob has 2 WETH
-	require.Equal(eth.Ether(2), dsl.Read(weth.BalanceOf(bob.Address())))
+	require.Equal(eth.Ether(2), contractdsl.Read(weth.BalanceOf(bob.Address())))
 }
