+++ op-devstack/dsl/contractdsl/call.go
@@ -1,11 +1,13 @@
-package dsl
+package contractdsl
 
 import (
 	"fmt"
 
+	"github.com/ethereum-optimism/optimism/op-devstack/dsl"
 	"github.com/ethereum-optimism/optimism/op-service/eth"
 	"github.com/ethereum-optimism/optimism/op-service/txintent"
 	"github.com/ethereum-optimism/optimism/op-service/txintent/bindings"
+	"github.com/ethereum-optimism/optimism/op-service/txintent/contractio"
 	"github.com/ethereum-optimism/optimism/op-service/txplan"
 	"github.com/ethereum/go-ethereum/core/types"
 )
@@ -28,16 +30,16 @@ func checkTestable[O any](call txintent.CallView[O]) TestCallView[O] {
 // Read executes a new message call without creating a transaction on the blockchain
 func Read[O any](call txintent.CallView[O], opts ...txplan.Option) O {
 	callTest := checkTestable(call)
-	o, err := txintent.Read(call, callTest.Test().Ctx(), opts...)
+	o, err := contractio.Read(call, callTest.Test().Ctx(), opts...)
 	callTest.Test().Require().NoError(err)
 	return o
 }
 
 // Write makes a user to write a tx by using the planned contract bindings
-func Write[O any](user *EOA, call txintent.CallView[O], opts ...txplan.Option) *types.Receipt {
+func Write[O any](user *dsl.EOA, call txintent.CallView[O], opts ...txplan.Option) *types.Receipt {
 	callTest := checkTestable(call)
 	finalOpts := txplan.Combine(user.Plan(), txplan.Combine(opts...))
-	o, err := user.Write(call, finalOpts)
+	o, err := contractio.Write(call, callTest.Test().Ctx(), finalOpts)
 	callTest.Test().Require().NoError(err)
 	return o
 }
