+++ op-devstack/dsl/eoa.go
@@ -98,12 +98,6 @@ func (u *EOA) Transact(opts ...txplan.Option) *txplan.PlannedTx {
 	return tx
 }
 
-// Write executes a tx by using the planned contract binding
-func (u *EOA) Write(call txintent.Call, opts ...txplan.Option) (*types.Receipt, error) {
-	finalOpts := txplan.Combine(u.Plan(), txplan.Combine(opts...))
-	return txintent.Write(call, u.ctx, finalOpts)
-}
-
 // balance looks up the user balance in the latest block.
 // It is not exposed publicly in DSL: see methods like VerifyBalance instead.
 func (u *EOA) balance() eth.ETH {
